Put wallets into own folder
