+++ .gitignore
@@ -20,7 +20,6 @@ node_modules/
 .DS_Store
 *.pem
 /*.sql
-lnbits/
 
 # debug
 npm-debug.log*
