+++ wallets/lightning-address/server.js
@@ -1,26 +1,5 @@
 import { lnAddrOptions } from '@/lib/lnurl'
-import { lnAddrAutowithdrawSchema } from '@/lib/validate'
 import { fetchLnAddrInvoice } from '@/lib/wallet'
-
-export const name = 'lightning-address'
-export const shortName = 'lnAddr'
-
-export const fields = [
-  {
-    name: 'address',
-    label: 'lightning address',
-    type: 'text',
-    autoComplete: 'off'
-  }
-]
-
-export const card = {
-  title: 'lightning address',
-  subtitle: 'autowithdraw to a lightning address',
-  badges: ['receive only', 'non-custodialish']
-}
-
-export const schema = lnAddrAutowithdrawSchema
 
 export const server = {
   walletType: 'LIGHTNING_ADDRESS',
