+++ wallets/lnc/client.js
@@ -1,35 +1,10 @@
-import LNC from '@lightninglabs/lnc-web'
-import { Mutex } from 'async-mutex'
-import { Form, PasswordInput, SubmitButton } from '@/components/form'
 import CancelButton from '@/components/cancel-button'
+import { Form, PasswordInput, SubmitButton } from '@/components/form'
 import { InvoiceCanceledError, InvoiceExpiredError } from '@/components/payment'
 import { bolt11Tags } from '@/lib/bolt11'
+import LNC from '@lightninglabs/lnc-web'
+import { Mutex } from 'async-mutex'
 import { Status } from 'wallets'
-import { lncSchema } from '@/lib/validate'
-
-export const name = 'lnc'
-
-export const fields = [
-  {
-    name: 'pairingPhrase',
-    label: 'pairing phrase',
-    type: 'password',
-    help: 'We only need permissions for the uri `/lnrpc.Lightning/SendPaymentSync`\n\nCreate a budgeted account with narrow permissions:\n\n```$ litcli accounts create --balance <budget>```\n\n```$ litcli sessions add --type custom --label <your label> --account_id <account_id> --uri /lnrpc.Lightning/SendPaymentSync```\n\nGrab the `pairing_secret_mnemonic` from the output and paste it here.'
-  },
-  {
-    name: 'password',
-    label: 'password',
-    type: 'password',
-    hint: 'encrypts your pairing phrase when stored locally',
-    optional: true
-  }
-]
-
-export const card = {
-  title: 'LNC',
-  subtitle: 'use Lightning Node Connect for LND payments',
-  badges: ['send only', 'non-custodialish', 'budgetable']
-}
 
 const XXX_DEFAULT_PASSWORD = 'password'
 
@@ -48,8 +23,6 @@ export async function validate ({ pairingPhrase, password }, { me, logger }) {
     lnc.disconnect()
   }
 }
-
-export const schema = lncSchema
 
 const mutex = new Mutex()
