+++ wallets/nwc/client.js
@@ -1,24 +1,5 @@
 import { parseNwcUrl } from '@/lib/url'
-import { nwcSchema } from '@/lib/validate'
 import { Relay, finalizeEvent, nip04 } from 'nostr-tools'
-
-export const name = 'nwc'
-
-export const fields = [
-  {
-    name: 'nwcUrl',
-    label: 'connection',
-    type: 'password'
-  }
-]
-
-export const card = {
-  title: 'NWC',
-  subtitle: 'use Nostr Wallet Connect for payments',
-  badges: ['send only', 'non-custodialish']
-}
-
-export const schema = nwcSchema
 
 export async function validate ({ nwcUrl }, { logger }) {
   const { relayUrl, walletPubkey } = parseNwcUrl(nwcUrl)
